api_project/pyproject.toml
[project]
name = "api-project"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "django>=5.2.8",
    "djangorestframework>=3.16.1",
]
